Refactor to mini-app that can easily be digested by different module
