+++ src/interfaces/http/modules/user/router.js
@@ -1,0 +1,200 @@
+const { Router } = require('express')
+const Status = require('http-status')
+
+module.exports = ({
+  getUseCase,
+  postUseCase,
+  putUseCase,
+  deleteUseCase,
+  logger,
+  response: { Success, Fail }
+}) => {
+  const router = Router()
+
+  /**
+ * @swagger
+ * definitions:
+ *   user:
+ *     properties:
+ *       id:
+ *         type: string
+ *         format: uuid
+ *       firstName:
+ *         type: string
+ *       lastName:
+ *         type: string
+ *       middleName:
+ *         type: string
+ *       email:
+ *         type: string
+ *       roleId:
+ *         type: number
+ *       isDeleted:
+ *         type: number
+ *       createdBy:
+ *         type: string
+ *         format: uuid
+ */
+
+  /**
+ * @swagger
+ * /users:
+ *   get:
+ *     tags:
+ *       - Users
+ *     description: Returns a list of users
+ *     security:
+ *       - JWT: []
+ *     responses:
+ *       200:
+ *         description: An array of users
+ *         schema:
+ *           type: array
+ *           items:
+ *             $ref: '#/definitions/user'
+ *       401:
+ *        $ref: '#/responses/Unauthorized'
+ */
+  router
+    .get('/', (req, res) => {
+      getUseCase
+        .all(req, res)
+        .then(data => {
+          res.status(Status.OK).json(Success(data))
+        })
+        .catch((error) => {
+          logger.error(error) // we still need to log every error for debugging
+          res.status(Status.BAD_REQUEST).json(
+            Fail(error.message))
+        })
+    })
+  /**
+ * @swagger
+ * /users:
+ *   post:
+ *     tags:
+ *       - Users
+ *     description: Create new user
+ *     security:
+ *       - JWT: []
+ *     produces:
+ *       - application/json
+ *     parameters:
+ *       - name: body
+ *         description: User's Entity
+ *         in: body
+ *         required: true
+ *         type: string
+ *         schema:
+ *           $ref: '#/definitions/user'
+ *     responses:
+ *       200:
+ *         description: Successfully Created
+ *         schema:
+ *           $ref: '#/definitions/user'
+ *       401:
+ *         $ref: '#/responses/Unauthorized'
+ *       400:
+ *         $ref: '#/responses/BadRequest'
+ */
+  router
+    .post('/', (req, res) => {
+      postUseCase
+        .create({ body: req.body })
+        .then(data => {
+          res.status(Status.OK).json(Success(data))
+        })
+        .catch((error) => {
+          logger.error(error) // we still need to log every error for debugging
+          res.status(Status.BAD_REQUEST).json(
+            Fail(error.message))
+        })
+    })
+  /**
+   * @swagger
+   * /users:
+   *   put:
+   *     tags:
+   *       - Users
+   *     description: Update User
+   *     security:
+   *       - JWT: []
+   *     produces:
+   *       - application/json
+   *     parameters:
+   *       - name: id
+   *         in: path
+   *         required: true
+   *         description: User's ID to update
+   *         type: string
+   *       - name: body
+   *         description: User's Entity
+   *         in: body
+   *         required: true
+   *         type: string
+   *         schema:
+   *           $ref: '#/definitions/user'
+   *     responses:
+   *       200:
+   *         description: Successfully Updated
+   *         schema:
+   *           $ref: '#/definitions/user'
+   *       401:
+   *         $ref: '#/responses/Unauthorized'
+   *       400:
+   *         $ref: '#/responses/BadRequest'
+   */
+  router
+    .put('/:id', (req, res) => {
+      putUseCase
+        .update({ id: req.params.id, body: req.body })
+        .then(data => {
+          res.status(Status.OK).json(Success(data))
+        })
+        .catch((error) => {
+          logger.error(error) // we still need to log every error for debugging
+          res.status(Status.BAD_REQUEST).json(
+            Fail(error.message))
+        })
+    })
+
+  /**
+   * @swagger
+   * /users:
+   *   delete:
+   *     tags:
+   *       - Users
+   *     description: Delete User
+   *     security:
+   *       - JWT: []
+   *     produces:
+   *       - application/json
+   *     parameters:
+   *       - name: id
+   *         in: path
+   *         required: true
+   *         description: User's ID to delete
+   *         type: string
+   *     responses:
+   *       200:
+   *         description: Successfully Deleted
+   *         schema:
+   *           $ref: '#/definitions/user'
+   *       401:
+   *         $ref: '#/responses/Unauthorized'
+   */
+  router
+    .delete('/:id', (req, res) => {
+      deleteUseCase
+        .remove({ id: req.params.id })
+        .then(data => {
+          res.status(Status.OK).json(Success(data))
+        })
+        .catch((error) => {
+          logger.error(error) // we still need to log every error for debugging
+          res.status(Status.BAD_REQUEST).json(
+            Fail(error.message))
+        })
+    })
+  return router
+}
